@@ -2,8 +2,7 @@
 // Created by libaoyu on 18-11-7.
 //
 
-#ifndef WINDOWOPTIMIZATION_TYPES_H
-#define WINDOWOPTIMIZATION_TYPES_H
+#pragma once
 
 //#define RES_DIM 3
 //#define FRAME_DIM 8
@@ -12,7 +11,3 @@
 //#define SCALAR double
 
 extern const float _G_gradStereoThershold;
-
-
-
-#endif //WINDOWOPTIMIZATION_TYPES_H
